app/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
cssutils==2.9.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
lxml==5.1.1
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pluggy==1.4.0
premailer==3.10.0
pydantic==2.6.4
pydantic_core==2.16.3
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
yagmail==0.15.293
